Consolidate text classes.
