--- src/flash/text/TextExtent.ts
@@ -1,38 +1,0 @@
-/**
- * Copyright 2014 Mozilla Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-// Class: TextExtent
-module Shumway.AVMX.AS.flash.text {
-  export class TextExtent extends ASObject {
-
-    constructor(width: number, height: number, textFieldWidth: number, textFieldHeight: number,
-                ascent: number, descent: number) {
-      super();
-      this.width = +width;
-      this.height = +height;
-      this.textFieldWidth = +textFieldWidth;
-      this.textFieldHeight = +textFieldHeight;
-      this.ascent = +ascent;
-      this.descent = +descent;
-    }
-
-    width: number;
-    height: number;
-    textFieldWidth: number;
-    textFieldHeight: number;
-    ascent: number;
-    descent: number;
-  }
-}
